[ADD] valeur de retour d'erreur builtins, [ADD] modification prompt en cs d'erreur de cmd precedente, [IN PROGRESS] execve cmd (msg d'erreur incomprehensible

+++ utils/str_utils.c
@@ -6,15 +6,15 @@
 /*   By: msharifi <msharifi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 14:32:11 by msharifi          #+#    #+#             */
-/*   Updated: 2022/11/22 17:41:24 by msharifi         ###   ########.fr       */
+/*   Updated: 2022/12/07 16:24:31 by msharifi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!str)
